Fix DNS seed initialisation
This fix updates the DNS seeds with proper names for Diamond. The DNS seeds are host names, which when queried return a list of IP addresses of active network nodes, that are then added to the list of nodes during initial startup and contacted for more network addresses. This only happens if no p2p nodes are configured (wallet already ran and has peers.dat, or addnode= or connect= are configured).
Some identation is also fixed in the file.