Remove extra feen from MultiSend
When MultiSend redirects an amount, it would remove 500 satoshi from the amount. This appears as fees, but is in fact not accounted for in the fee amount. It is just 500 satoshi less.

Other transactions do not have such amount substracted (except the questionable AutoCombineDust() routine which if not fixed, should be removed) -- normal fee rules are applied.

This leaves questionable line 3530, which checks for 500 satoshi as well -- but without modifying that like the fix still works - probably that should be looked at as well.